refactor: stations를 일급컬렉션으로 변환
